Move ProgressBarMixin and better readability
