+++ management/commands/normalize_partnerships_supervisors.py
@@ -1,6 +1,6 @@
 from django.core.management import BaseCommand
 
-from partnership.management.commands import ProgressBarMixin
+from partnership.management.commands.progress_bar import ProgressBarMixin
 from partnership.models import Partnership, UCLManagementEntity
